@@ -8,8 +8,9 @@ use Livewire\Component;
 
 class CommentsSection extends Component
 {
-    public $post;
+    public Post $post;
     public $comment;
+
     public $successMessage;
 
     protected $rules = [
@@ -27,6 +28,7 @@ class CommentsSection extends Component
         $this->validate();
 
         sleep(1);
+
         Comment::create([
             'post_id' => $this->post->id,
             'username' => 'Guest',
@@ -35,8 +37,12 @@ class CommentsSection extends Component
 
         $this->comment = '';
 
+        //force livewire to render and show the new post due to state change
         $this->post = Post::find($this->post->id);
 
+        //You can also use $this->post->refresh() to re-hydrate the Post model using fresh data from the database.
+
+        //session()->flash('success_message','Comment was posted');
         $this->successMessage =  'Comment was posted!';
     }
 
